some changes related redirecting

+++ src/main/java/com/crudapp/Crud/App/service/UserServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -27,23 +28,26 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUserbyUsername(String username) {
+    public User getUserbyUsername(String username){
         return userRepo.findByUsername(username);
     }
 
     @Override
     public String deleteUser(String id) {
-        userRepo.deleteById(id);
-        return "User deleted successfully";
+        if (userRepo.existsById(id)) {
+            userRepo.deleteById(id);
+            return "User deleted successfully";
+        } else {
+            return "User does not exist";
+        }
     }
 
     @Override
     public String updateUser(User user) {
-        if(userRepo.existsById(user.getId())){
+        if (userRepo.existsById(user.getId())) {
             userRepo.save(user);
             return "User updated successfully with username " + user.getUsername();
-        }else
-            return "User does not exist";
+        } else return "User does not exist";
     }
